Changed navigation / level data handling
Changed the way the navigation system is used. The old system was using too many files, so I copied the code and placed it into a new file- Navigation.js. This will allow me to easily see all the code that is used when the user moves through the game to different screens / levels.

Fixed bugs where the number of stars and node coords were not correctly being stored.

Started work on a way to distinguish between whether the user is in the sandbox, or whether he is in a level. My idea is to make a variable inSandbox, that changes depending on whether he is in the sandbox or not.