The level selector works! With a few bugs
The user can now select a level from the campaign menu. I have created a temporary file called Visualizer2 which houses some functions that both sandbox and the level selector will use.

There is a slight bug however, I disabled the option to drag the start and end nodes in the level, but when the node is created, it will still have the mouse event listeners- and therefore will cause an error if the user trys to click  or drag ect.