Fixed previous error -- 2 new bug
Removed the grid constant and just used the currentGrid. The dijkstra algorithm, along with the animations work correctly!

The sandbox mode also works. I do need to tidy it up and place it all in a separate file and remove the old visualizer.

The new bug-- if there is no path, on a level,  the error animation is a bit wonky. The distances still appear on the screen, and the animation itself is not performing correctly. This is a logic error- as there is no actual errors.

Second bug-- the level grid is no resetting every time the user goes back to the home screen. The walls are still there - although the animation is not...