Still has bug from before - one more bug
The bug from before is still here.

Removed the old visualizer all together. Sandbox and Campaign now work separately.

New bug: load grid is not working on the sandbox. It says that it is not a function. I think I forgot to get rid of the "this" keyword on the function call