fixed previous big bug new bug-- error on starting dijjkstra
fixed previous big bug.

I was changing the properties of the variable. I now make copies of the original variables and modify those and reset them each time.

Also fixed an issue where the variables of the current level were not being reassigned every time the level was entered, so the current level data did not change.

There is also now a bug where the start node of level 2 does not work correctly and there is an error running the actual dijkstra algorithm.