Fixed bugs and added a few things
Fixed a bug where the when loading a test grid, the coords of the start  or end nodes where not being set, so even though it looked like the start/end nodes were in the middle, the actual coords of them (that were passed into the dijkstra algorithm) were not being changed accordingly.
Also changed made it so if the nodes were reset during animation, the distances were set to a blank space, so it did not look weird when animating the rest of the nodes. (blank space from 'Infinity')