Remade level 2 and changed random wall button appearance
Recreated level 2 to be complete. Changed its name in the json file.

Also changed when the numRandomWallPresses (left) text message is displayed: It is only displayed if the ORIGINAL value is not 0.

At the same time, I modified when the randomWallButton is shown. Previously, it was showing when the CURRENT numRandomWallPresses (left) was 0, now it still shows if that value is 0- to remind the user that they have no more random wall presses left. This is different to showing when the ORIGINAL value is 0.