Changed many things and fixed a few bugs
The main thing added in this update is the ability to add a permanent wall into a level. There is a now a toggle permanent wall button on the sandbox level. If toggled, any wall you place will turn into a permanent wall and if you click on an existing wall, it will be turned into a permanent wall. These permanent walls cannot be removed in the campaign levels. permanent walls have the same properties as normal wallls, except they have different css which means I did not need to change much existing code. I did have to consider the walls used text on the sandbox mode. I also had to change resetAllNodes to make sure the permanent walls keep their classes.

I added a test level with a few permanent walls and they get implemented correctly.

For the actual toggle node button, I basically created a copy of the standard node class with a few different features.

I also changed a bit of css of the actual  levels/sandbox so text gets rendered correctly.

Fixed a bug where the user could untoggle already placed walls on levels, even if the number of active walls is 0. This has been fixed by adding a few more conditions on the toggle wall if statement. The user can now no longer turn a wall into a non wall if the number of active walls is 0.