Changed css, level name is displayed
Changed some css for some elements. This includes removing margins and therefore changing the fixed positions of some elements.

The name of the level the user is currently on is displayed in the top left corner. I used percentages for the right and left padding which creates a small border around the level name that changes depending on the length of the level name so that if the level name is very small, it does not look weird.