Started to add a proper home screen (not title)
I started to add a simple home screen which will allow the player to choose either campaign, which will take them to the main title screen, or the simple sandbox mode.

It consists of a simple light affect made using CSS.

The bulb does scale up and down well enough, but sometimes the actual light affect does not look smooth at all.

Known bugs:
Errors thrown when user tries to leave the "arena" when an animation is playing. Need a way to cancel the animation when this happens.

The grid itself in the "arena" does not scale up or down well at all. This needs to be fixed as it is a problem for screens of other dimensions.