Started work on a progress bar for the home screen
I have started to implement a loading bar for the homescreen when the user tries to enter the campaign. Its only function is to look cool.

The aim is to have the loading bar fill to 100% and then the user enters the main menu, aka the title screen.

Currently, it does not really work yet, as there is a problem with displaying the bar percentage.