Fixed big bugs.
Fixed a big bug that was annoying me for a long time.

Up until now, if the user tried to enter the home screen, errors would be displayed everywhere. Now, a class is added to the home button, "enabled". If enabled, the user can leave to the home, else, nothing happens.

The user can leave to the home screen when all animations are finished.

Also fixed the loading bar entirely. Now, the user can go to both the campaign and the sandbox.

New Bugs:

The new fix above is not permanent and hopefully a proper fix will be made later- especially with a pause button.

This fix currently only works on the Sandbox mode. Errors are still thrown on the actual campaign. This should be a quick fix as I just need to add a "homeButton" id to the home button in the level visualizer- and it should work the same.

The user can also make similar bugs occur if the user starts the algorithm again both on the sandbox and on the level. Therefore this fix needs to be applied to all the buttons on the levels- especially the campaign.