Started work winning/losing features
Started work on experimenting with animating when the "missile" travels a certain distance.

For example, if the distance the shortest path "missile" has travelled is further than 70, it "explodes", leaving a head and a body of the missile trail.

Currently, there are a few issues. I need a way of the missile "ending" only once.

I want the "ending" animation to play from reverse once, staring at the head.

I have already added the head and body css for the missile ending.