EndTrail function fixed
The main statements in the function endTrail can only run once, no matter how many times the function is called.

This ensures that the head and body of the "ended" missile are only animated once.

It works by calling an empty function once the function has run once, by setting a flag.