Fixed bugs
Fixed a previous bug where if the missile reached the end node, the shortest path would not be animated.

I fixed this by setting the initial value of endIndex to infinity.

Also fixed a bug where the user could leave to the home screen while the missile ending animation was playing.
This was caused because the check that the animation was playing was using the reversed variable x, so it started of at 0, so the program thought the animation ended as soon as it begun.

I can to change the check that x was > than endIndex -2