Fixed bug
Fixed a bug where the missile ending animation would not play immediately after the missile ends.

This was due to the program still looping through the rest of the shortest path, even though it would not animate them. This was not noticeable in many levels, as the shortest path was not that long, but when the shortest path was very long, it took a long time for the end animation to play, which was how the bug was found.

I fixed this bug by creating another variable called looplength, which specified how many times the main for loop repeats. Before the change, the loop was set to repeat at the number of nodes in the shortest path. It now has two options on how many times to loop.

1. The missile reaches the end node. This means that the entire shortest path list should be animated- and the variable is set to the length of the shortest node path array.

2. The missile ends before it reaches the end node. In this case, the variable is set to the number of nodes before before the missile ends.

You set the variable by checking whether the missile will end, before it has ended. If the endIndex has not changed, then the missile does not end.