Massive Updates
Huge updates:

This was the first major step forward with improving the game, and most importantly, looking ahead.

The first big change was increase the number of rows from 21 to 23. And reduce the number of columns from 51 to 49. This was done in an attempt to help a big scaling issue which has gone unnoticed up until now. I realised that the resolution of the game window was not the default resolution- and thus all past changes to scaling and element placing have been scaled to this random zoom resolution. At the "actual size", I found that pretty much every element needed re-placing. The grid also did not fit onto the screen, thus I had to remove some of the columns.

I re-placed all the elements to their proper positions. I also had to change the start and end default coords on the sandbox mode to those of the new grid size.

I also had to redo all the levels. Currently, I have only redone the Symmetrical Surveillance and Every Other One, the two better levels in my opinion. I also changed the order of the levels, and set Every Other One to level 1. I have yet to redo level 3, the Three Walled Rows, and the two test levels- as these are not critical to redo.

I first got an error when trying the newly redone levels and remembered I did not change the start and end coords in the json files for the level data to the new grid sizes. Once these had been changed, it all worked fine.

Apart from this minor error, there were no other complications.

I changed all the walls on Symmetrical Surveillance to permanent walls, just to test that they still work, and they do. I need to add permanent walls to the other levels, but that is a later thing to do.

The next stage is to get a working campaign screen with locked levels and level progressing.