Added fonts and changed overflow settings
Added 5 different fonts that I could use for text in my game. They are otf and ttf formats, which can be referenced with css.

I also made it so any overflowing elements in the html page will be hidden, which effectively removes the vertical scroller appearing. This will override any css styles. This makes it easier for me to layer elements on the pages, as even if the elements go slightly out of the page, it won't mess anything up.