Added more levels and updated the levelDataManager
I have added 15 levels in total. Only 1, 2 and 3 are actual levels currently, the other 12 are just test levels with the same data. There will be a total of 15 campaign levels in the game, and since the program renders these buttons, it allows me to decide how the campaign menu will look, as it needs to comfortably fit these 15 buttons to each of the levels.

I also added two more values to each level, a difficulty and a description. These will be rendered in the campaign menu.

I also had to change the allLevelData file a bit to parse the level data accurately with these extra changes.

I added functions to get all of the difficulties for the levels and the descriptions. I also added functions to get individual info about a level, given the ID of the level.

It did not work at first because I forgot to change the pre-set value of the number of levels from 5 to 15, which meant only the first 5 levels were being rendered.