Added homescreen background, testing background, and changed css
As the title suggests, I added both a main background which I am planning to use as a background for all major menus in the game, such as the home screen and the campaign screen.

I also added a testing background which is a bright red color to help me easily see where the white screen begins and ends in my program, to help me layer items accordingly.

I also changed the grid position in sandbox and changed the border radius to make the pink container a square- instead of having rounded edges.

The entire game UI is going to change later, but it is more important to have a menu and campaign, so I can match the theme to the game window.

I will redo all the levels and change the number of nodes in both the rows and columns, when I know the size of the other elements that will be placed in the screen.