Added a separate way to go home from a menu
When I create the campaign menu, or any other menu, I will need a separate function to take the user back to the homepage. This is because, the original home navigation function only allowed the user enter to the homescreen, if an animation was not playing, which required checking a class on a button in the game screen. This button is not being rendered when the user is not in the game screen, so it would cause an error. The new function will allow the user to enter home with no conditions needed to be met.