Finished campaign menu v1
I have tidied up and added to the campaign menu.

I have added lives and allowed walls to the level info container. I also added tags to each container, to tell the user what the text/number represents (eg. Difficulty --- Easy)

I still need to add the thumbnail images- but I cannot do this without creating the completed levels and creating the UIs for the actual level screen. I have left a space in the level info container for which these thumbnails will go.

There was a slight bug which meant the user could highlight the text displayed on the screen. I fixed this by adding a css line in the body tag, which will effect every element rendered, 'user-select: none'. This stops the user being able to select the text.

I also tidied up the css by creating a default css class for each of the tags and text and containers inside the level info container (except the description one, because that needs a different width/color/height than the others). This means I only need to change the top and left values of the elements, and I don't need to copy and paste as much code.