Started work on in-game UI
I have started to add UI into the in-game screen.

I have added a top divider, which will house the buttons-just like the home screen.

I have also added a "Start" button to run the actual algorithm.

I introduced a max character length of 22 in the username input box in the home screen.

I have added a background colour in the in-game screen- and changed the width and height of the nodes to fit the screen- as well as change the outline of the nodes to match the screen.

I have also increased the number of rows and columns of nodes.

In the campaign screen, I fixed an issue where the level names after level 9 where further to the right by a few pixels than names 1-9, because of the size of the text of the level id. 1-9 is smaller than 10-15. I had to change the margin size for these level names only.