Added local storage for userName and the number of levels unlocked
As title suggests, when the user enters a name in the userName input field, and then enters campaign or sandbox, their username is stored to local storage (if it hasnt been stored yet), or updated.

The username will also be displayed in the input box when the user goes back into the homescreen (either already running, reloaded, fresh start).

There is also a check on the username. It does not accept special characters (including spaces) inside the name and it cannot be a blank. If there are these things present in the name, an error message in the input field is displayed in red text, and then removed after 3000 ticks. If the user clicks on the input box before this time, it gets manually removed and the color set back to white.

The number of levels the user has unlocked is also stored the same way and automatically set to one, or gotten.

Currently, this value is fixed because there is no winning/losing present.