Added error checking. Last level unlocked is automatically selected
I have implemented an error checking system for the username input box.

If the username entered is a blank space, or equals "ERROR: INVALID INPUT" (more on that later), includes a blank space, or contains a special character, the value of the input box gets set to "ERROR: INVALID INPUT" in a red colour.

After 3000 ticks, the input box field is set to empty, and the colour is set back to white. However, if the user clicks the box before the timer runs out, the value is automatically set to empty, and the colour white.

There was a slight bug where if the user clicks the box before the timer runs out, the changes happen, but then if the user types a proper name, the changes happen again (after a few seconds)! Why? Because the original timer (3000 ticks), does not stop, and so when it runs out, the changes happen, whether or they already happened! To fix this, I set a check in the timer loop. When the timer runs out, the changes only occur if the current value of the input box equals the error message (ERROR: INVALID INPUT). This means if the user has changed the value to "Bob" (for example), when the timer runs out, the changes do not occur for a second time.

I also made it so that the last level that the user has unlocked is automatically clicked, and the level info for that level is displayed. I did this by removing the "first" variable (as before this update, the level that automatically was clicked was always the first level).  I replaced this variable with "lastLevelUnlocked", so even if the last level the user has unlocked is the first level, or the last, it doesnt matter.

There was also a slight bug with this as well. At first, the right level was being selected automatically, but the level info for this level was not being displayed. The level info being displayed was that for the first level unlocked. It turns out that I had not made the default value of levelClicked equal to the numLevelsUnlocked. previously, this value was 1, which is why the first level was being shown in the level info. Once I did this, it worked.