Started work on Options menu
Changed home screen Settings button to Medals. Therefore have to change positioning of the elements.

Started work on Options menu. Added a state showOptionsMenu, set to False as default. The Options button changes whatever the boolean value is of this state variable. This re-renders and the optionsMenu component is returned by getOptionsMenu. This creates a div that covers the whole screen that blurs its background and adds a light blackish colour. The modal window has a Back button which hides the options menu