Bug with permanent wall toggling
I have to change the way I handle permanent walls being toggled. I have created a global variable boolean permanentWallToggled. The variable is correctly being toggled, and the class should be being set, but it isnt and is reverting back to its original state each time the menu is opened.