Setting current end distance now works
I created a new function saveOptions() which will eventually save the current settings the user has set. Currently, only the endDistance can be saved - and thus fixes the bug where the game could not be ran as the value of the input box was trying to be parsed when it didnt exist because it was in the options menu.

There is still the bug of the grid checkbox reverting back to its original form when toggled and the menu is exited and re-entered