Re-worked options menu and other clickable nodes
Re-named "NodeClickable" component to NodeChangeWallType to help aid readability.

I have decided that the current checkbox element that the user can click on is too complicated to work with. I have therefore updated the Permanent Wall Holder graphics to fit in with the rest of the options menu. I then used this template to make a copy to hold the toggle grid button, which is the same graphics as the permanent wall graphics. This will make it much easier to modify the state of it. I therefore had to create a new NodeToggleGrid component, and add some extra code in the toggleGrid() function to change the class of the clickable node to toggle its colour. This all works perfectly, except the ToggleGrid outline has a lot of padding, which is a bug.

I have also added a Development Options sub-container to hold the Set End Distance input box, which now has some text to explain what it is. The saveGrid and loadGrid buttons are also here as well.