Toggle Grid now saves its state
Fixed previous bug where the Toggle Grid button did not save its current state each time the menu was re-entered.

I removed the gridOn boolean state, as if I used it to track the current state of the grid outline, all animations would be re-rendered, which isn't good.

I therefore created a  new global variable which I can toggle, similar to the permanentWallToggle. I then call the toggle grid outline function each time toggleGrid() is called. I pass the state of the global variable into the props of NodeToggleGrid as the curretstate to display to the user.