Error due to plane animation when trying to enter home
Huge error where if you use the Random Wall Button, there is an error if you try to go home, as the plane animation is still trying to animate. You can also go home while the animation is playing, which of course throws an error.

Fixed first issue by adding a condition where if the animation has ended, clear the interval

Fixed last issue by parsing in if the animation is playing on the EnterHome, if it is playing, then stop the person from leaving.