Started re-working error handling animations
The error message gets correctly displayed, but I want the "ERROR:" part of the message to be in a different colour. Luckily, I made that part out of permanent walls, so if the current node is a permanent wall, it is important and should be appended to the importantNodes array and have a separate animation. This does not happen, and these nodes are animated along with the rest of the message in the same way.