Re-worked level visualizer
Created level 1.

Added relevant sandbox code (so not including development features, such as permanent walls, or saving grids, or dragging nodes, or always allowing permanent wall)

The user can now enter the first level, and the grid is correctly displayed with the correct graphics.

The error message (system not secure) is correctly displayed if there is not a path.

The random wall button only displays if the user can use it.

Options menu only has non-development stuff.

Updated graphics for the message showing how many walls the user has used

There currently is a bug where the shortest path animation does not reach the end node, but is stopping at 69 distance. This is probably due to some old code from before.