Added end distances in levels
Added new functions to correctly get the end distance for each level, which can be put in the json files. If the user is in Sandbox, then the end distance that the user sets is used. If the user is in a level, then the pre-set value is used. The algorithm correctly animates the "end trail" animation at the given distance.