Errors and fixed bug
Added conditional if statements to stop the user clicking the Run button while the animation is playing. (fixing bug)

The number of levels the user has unlocked is default 1 if it is the first time running. Else, the number of levels is set to the stored value in local storage.

Error thrown when trying to enter level without specifically clicking on it. The level info is shown.

The numLevels unlocked is not incrementing by one when the trail ends and the current level is the same as the number of levels unlocked, even though the function is running once.