Number of random wall presses remaining fixed
Moved the number of walls used to the right side of the screen.

The text that displays the number of random wall presses left is on the left side, and correctly counts down till 0. When it is 0, the button Random Walls gets 'disabled' and blacks out so the user cannot click it anymore.

There seems to be an issue that happens only in the level. When the random wall button is pressed, the last 5 or so columns do not get updated straight away with their walls, and they only get updated at the end when the plane has finished flying. This only occurs on the level, not the sandbox