Fixed random wall grid not updating on level visualiser
As title suggests, the grid now gets updated on the last iteration of the walls being added.

The user cannot run the Random Wall animation, if the dijkstra is already running on the level screen now, fixing the bug.

The user can however start the dijkstra algorithm if the random wall animation is playing, which is a bug on both screens.