Started work on dialogue manager
Started work on new dialogue manager. The dialogue is stored in a json file in a 2d array. A '1' means that it is the user speaking. '2' means the next main character. '3' means just sound effects or other things that nobody says.

Currently, each new line of dialogue can be got by clicking the button in the development menu, which means that I can now parse each line, checking for any variables in the string (and replacing them), and then eventually return everything to dialogue menu to display to the user. The user will see each line of dialogue (or sound effect) when they hit the Next button, or enter key.