The dialogue menu now opens
If the tutorial menu has finished (ie the user has clicked out of it), then the dialogue menu can be shown.

There is still the bug where if the tutorial menu doesn't normally get shown (for levels >1), then the dialogue menu will never be shown. I need to add a check for this.