Third attempt - dialogue menu
The shouldChange is correctly working (from previous commit), but the line number is not updating

I had to update the currentDialogueLine before the text was parsed. This solved the issue of the line count not updating.

The results are correctly received, but the screen is not updating with them for some reason.