Fixed dialogue meu
This is the successful attempt at fixing the dialogue menu. The problem was that the state was never changing, so nothing was being re-rendered.

I now need to make it so the user will exit the menu when the last dialogue peace has been said, at the moment the program throws an error when this happens, which is expected because that part has not been written yet.

I still need to make each dialogue box go to the left if it is from another person, or on the right if it is the user speaking. I also need to keep the dialogue boxes on the screen, and not removing them each time.

Bug: options menu can be entered if the tutorial/dialogue menu is opened