Second fix attempt - dialogue menu closing
This is my second attempt at fixing the error. This time, I return two separate buttons if the user should exit on the next click, or not.

This time, everything still works, and no errors are thrown, but the menu does not close. This is probably an issue with my boolean variables that dictate whether the menu can open or not.