Fixed dialogue menu closing errors
I found that the closeDialogueMenu() function was indeed running, which confirmed my suspicion that it was to do with my boolean variables.

I created a new function called setHasShownDialogueMenu(), which allows me to set the specific value of hasShownDialogueMenu, so no need for toggle.

I also corrected (even though no errors were thrown/bugs found) the code for toggling whether the tutorial has ended. The value should not be toggled if it is True, only from going False to True should it be toggled.

I also setState manually to change showDialogueMenu, instead of the toggle.

All this seems to have worked, and the menu now closes on the last click.