Dialogue menu doesn't play again if re-entered from home
Even though the tutorial does not play again if you re-enter the level (as it shouldn't), the dialogue menu does not play again. I want the dialogue to be played every time a level is played, whether or not they have seen it before.

I fixed this by re-setting the dialogue variables and now works.