Started to re-write dialogue menu
I thought the way I had written the dialogue menu code was not very good, and would make it hard to implement the features I want to, specifically making it so previous messages were still visible.

I have changed it so a whole array is returned from the dialogue manager.

This array is used in getDialogueBlocks which pushes all the dialogue that should be visible to an array, returns this array. This array is then rendered.

Sound effects and other things are in italics.

The dialogue system now technically works. Each message is added, and then at the end, the user can exit the menu.

Although there are no "bugs," there are still some things I would like to do with the dialogue:

1. Have whoever-is-speaking-to-the-user 's text on the left side of the screen and have the user's text on the right hand side.

2. Have sound effects/other things appear either in the middle, or whoever it concerns

3. Place the speaker's text in the format <{speaker}>.

4. Fix the container width that contains the dialogue, its a bit tight

5. Colour code slightly the different people

I can also see some potential problems that will be dealt with nearer the time. The main issue will be when the dialogue messages have no room left. I will need to create a new page with back button.