Fixed dialogue menu
I have not started implementing the things mentioned in the last commit.

Fixed bug where the proper dialogue was not showing. This was because I was not importing my variables correctly in the dialogue manager. I also made sure that the currentLevel variable was always a number.

Fixed bug where if the dialogue had been played on a level > 1, and then the user went home and played the "tutorial" level, the dialogue menu would be on top of the tutorial menu. Fixed this by setting hasTutorialEnded to false each time the user goes to the home page.