Started work on dialogue visual features
Each dialogue line is given a 1 or 2 position in the json file. A 1 means that the text should be displayed on the left hand side of the dialogue menu, a 2 means the right hand side.

I removed the background of each text, as it didn't look right.

I still want to have the speaker's names a different shade of colour.

There will still be issues if the dialogue goes off the screen. Before I work on that, I will need to add the dialogue for all the levels. I also want to increase the width of the dialogue menu. I will do that first.