Continued work on dialogue menu
Added new features to the dialogue menu.

Removed the variable currentLineNumberEnd, because it was causing issues with the currentlevel initially being set to 1. The function is now always used, to make sure all values are up to date.

The words <hit enter> still get displayed below the last line, unless it is at the end, where <press enter> is displayed, or before a scene break, which <press next> is displayed. This means that the button Next is not always displayed.

A scene breaker can be added using the code 999, and the position being set to 3 (centre).

The code for if the dialogue is too big to fit in the menu still needs to be written.