Started work on next page function dialogue menu
I need a way to make it so the user can go to the next page of the dialogue meu, when there is not enough room in the menu.

Currently, the text <press continue>, is displayed when this should occur, designated by a 1000 in the json file. The button Continue is also displayed, and the user cannot hit Enter when this is showing.

My first solution was to just start looping i in Visualise Level from the currentDialogueLine number (if the next page should be on), so the previous blocks would not be rendered. Unfortuantly, this did not work