Attempt 2 dialogue next page
This attempt was a lot more successful. I scrapped the previous code, and this time added a new state variable dialogueStartLoop. When a new page is needed, this gets updated to the current dialogue line number.

This partially works, as the new page is correctly shown, except the previous line is still showing before the new lines!