Added story dialogue for all level and other things
The main thing in this commit is that the story mode dialogue is now complete.

Changed Powerups to Random Walls, as that is the only powerup I will have on campaign screen

Replaced Lives with End Distance on campaign screen.

I decided against doing a Medals/achievments menu because I simply don't have enough things I could award them for, so it would be a waste of time creating it.

<Mr Smith> (main protagonist) is now in a light red colour.

Added a new input box in the Sandbox mode which allows the user to load any level 1-15.

Obviously added many more things to replace in the Dialogue manager. Nothing else.

There is a tiny visual glitch where some text is appearing just slightly outside the screen.