Fixed bug with random walls on sandbox mode
The unWallable variable was not being set to True for permanent walls on the Sandbox mode. This meant that the dijsktra algorithm was just ignoring them when playing its animation cycle.

Increased opacity of credits text, as it was too faint.

(Don't worry about the levels.json, they are not ready)